server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
ultralytics==8.3.23
Pillow==10.4.0
PyYAML==6.0.2

